Fix issues after code review
