+++ src/__test__/app/api/todos/UpdateTodoController.spec.ts
@@ -1,30 +1,18 @@
 import { StatusCodes } from "http-status-codes";
-import jwt from "jsonwebtoken";
 
 import { PrismaClient } from "@prisma/client";
 
 import { TodoRepository } from "../../../../repositories/TodoRepository";
-import { UserRepository } from "../../../../repositories/UserRepository";
-import { requestAPIWithAuth } from "../../../helper/requestAPIWithAuth";
+import { requestAPIWithAuth } from "../../../helper/requestHelpers/requestAPIWithAuth";
+import { createTestUser } from "../../../helper/requestHelpers/requestAuthHelper";
 
 const prisma = new PrismaClient();
 
 describe("【APIテスト】Todo一件の更新", () => {
   let cookie: string;
+
   beforeAll(async () => {
-    const repository = new UserRepository();
-    const userData = await repository.register({
-      name: "ダミーユーザー",
-      password: "dammyPassword",
-      email: "dammyData@mail.com",
-    });
-    const userId = userData.user.id;
-
-    const token = jwt.sign({ userId }, process.env.JWT_SECRET!, {
-      expiresIn: "1h",
-    });
-
-    cookie = `token=${token}`;
+    cookie = await createTestUser();
   });
   describe("【成功パターン】", () => {
     beforeEach(async () => {
@@ -111,7 +99,6 @@ describe("【APIテスト】Todo一件の更新", () => {
       expect(response.body).toEqual({
         message: "入力内容が不適切(文字列のみ)です。",
       });
-      expect(response.statusCode).toEqual(StatusCodes.BAD_REQUEST);
     });
     it("ボディに不適切な値(文字列ではない値)が入力された場合、リクエストはエラーになる。", async () => {
       const data = 123;
@@ -126,7 +113,6 @@ describe("【APIテスト】Todo一件の更新", () => {
       expect(response.body).toEqual({
         message: "入力内容が不適切(文字列のみ)です。",
       });
-      expect(response.statusCode).toEqual(StatusCodes.BAD_REQUEST);
     });
     it("存在しないIDへのリクエストはエラーになる。", async () => {
       const response = await requestAPIWithAuth({
@@ -139,7 +125,6 @@ describe("【APIテスト】Todo一件の更新", () => {
       expect(response.body).toEqual({
         message: "Todoの更新に失敗しました。",
       });
-      expect(response.statusCode).toEqual(StatusCodes.NOT_FOUND);
     });
     it("プログラムの意図しないエラー(サーバー側の問題等)は、エラーメッセージ(InternalServerError)とstatus(InternalServerError=500)が返る。", async () => {
       jest.spyOn(TodoRepository.prototype, "update").mockImplementation(() => {
@@ -154,7 +139,6 @@ describe("【APIテスト】Todo一件の更新", () => {
       });
 
       expect(response.body).toEqual({ message: "Internal Server Error" });
-      expect(response.statusCode).toEqual(StatusCodes.INTERNAL_SERVER_ERROR);
     });
   });
 });
