+++ src/__test__/helper/requestHelpers/requestAPIWithAuth.ts
@@ -1,6 +1,6 @@
 import request from "supertest";
 
-import app from "../../app";
+import app from "../../../app";
 
 export const requestAPIWithAuth = ({
   method,
